Control.State.Transition.Invalid.Trace

Description

Invalid transition system traces.

An invalid trace consists of an valid prefix, and a last signal that might be invalid. The validity of the signal depends on the probability of the trace generators of generating invalid signals.

Documentation

data Trace s

Constructors

Trace

Fields

validPrefix :: !(Trace s)

signal :: !(Signal s)

Last signal in the trace. This might cause a predicate failure, but it isn't guaranteed to do so, since invalid trace generation is probabilistic.

errorOrLastState :: !(Either [PredicateFailure s] (State s))